# openapi.yaml
openapi: 3.1.0
info:
  title: Scrutineer Skill API
  version: 0.4.0
  description: |
    A narrow HTTP surface that claude-code skills use while they run. The
    goal is enough to let a skill orchestrate other skills and read prior
    scan outputs without pushing that logic into scrutineer's Go code.

    Every scrutineer skill scan is issued a bearer token that is staged into
    the workspace context.json. Skills authenticate with that token. A token
    is only valid while its originating scan is running, and only grants
    access to the repository the scan was issued against.

    The `/v1/*` paths are an unauthenticated, localhost-only export surface
    intended for external scripts (jq, cron, CI dumps). They emit JSONL and
    rely on the same host-header check as the browser UI; see threatmodel.md.
servers:
  - url: http://127.0.0.1:8080/api
    description: Local scrutineer server
security:
  - bearerAuth: []
paths:
...
  /repositories/{id}/findings:
    get:
      summary: List findings across every scan on this repository
      operationId: listRepositoryFindings
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
        - name: skill
          in: query
          description: Only return findings produced by scans of this skill (e.g. `semgrep`, `security-deep-dive`).
          schema: { type: string }
        - name: scan_group
          in: query
          description: Only return findings from scans in this parallel batch, so an in-flight audit skill reads only its siblings' findings.
          schema: { type: string }
        - name: sub_path
          in: query
          description: Only return findings for this monorepo sub-package (e.g. `activesupport`), so a repository can be exported or bundled one sub-package at a time.
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingSummary' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Stream one finding into the concurrent-finding log
      description: >
        Records a single finding emitted mid-scan so sibling scans in the same
        `scan_group` can read it before this scan completes. The body is one
        finding in the same shape it has in the skill's `report.json`; the
        scan's identity (scan id, repository, commit, sub-path) is stamped from
        the bearer token, never trusted from the body. The finding still
        belongs in the final `report.json`, which reconciles against the
        streamed row by fingerprint rather than duplicating it.
      operationId: streamRepositoryFinding
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FindingInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingSummary' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /findings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: Fetch one finding with its six-step prose
      operationId: getFinding
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update mutable fields on a finding
      operationId: patchFinding
      description: |
        The authenticated scan must be finding-scoped to this finding ID.
        Each field value is written through the history-tracked path.
        Finding-scoped skills write with source=model_suggested; browser
        form edits use separate routes and write with source=analyst.
        Unknown field names are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fields]
              properties:
                by:
                  type: string
                  description: Optional author string for the history row.
                fields:
                  type: object
                  additionalProperties: { type: string }
                  description: |
                    Map of field name to new value. Allowed fields:
                    title, severity, status, cwe, location, affected, reachability, quality_tier,
                    cve_id, ghsa_id, cvss_vector, cvss_v4_vector, fix_version, fix_commit,
                    resolution, disclosure_draft, disclosure_title, suggested_recipients, assignee,
                    suggested_fix, suggested_fix_commit, breaking_change, breaking_change_rationale,
                    exploited_in_wild, exploited_in_wild_evidence, mitigation, mitigation_semgrep,
                    release_tag, release_url, last_revalidate_verdict.
      responses:
        '204': { description: Updated }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/findings:
    get:
      summary: Export findings across all repositories as JSONL
      description: |
        Returns one JSON object per line (`application/x-ndjson`). No bearer
        token; localhost-only. Each line is a `FindingExport` carrying the
        full set of `db.Finding` columns.
      operationId: exportFindings
      security: []
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [jsonl] }
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
      responses:
        '200':
          description: JSONL stream of FindingExport objects, one per line.
          content:
            application/x-ndjson:
              schema: { $ref: '#/components/schemas/FindingExport' }
        '400':
          description: Unsupported format query value, or `encrypt`/`scope`/`include` requested (`scope` is only valid on per-repository exports, `encrypt`/`include` on per-repository bundle exports).
        '403':
          description: Host header is not localhost
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RepositoryID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
...
    Forbidden:
      description: Token does not grant access to this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Malformed query or request body, or a missing required field
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    FindingInput:
      type: object
      description: >
        One finding as it appears in a skill's report.json, streamed into the
        concurrent-finding log. Only title, severity and location are required;
        the other report.json finding fields (cwe, sinks, dup_check, trace, …)
        are accepted and stored. Scan-owned identity fields are ignored.
      required: [title, severity, location]
      additionalProperties: true
      properties:
        title:    { type: string }
        severity: { type: string, enum: [Critical, High, Medium, Low] }
        location: { type: string }
        cwe:      { type: string }
        sinks:    { type: array, items: { type: string } }
        dup_check: { type: string, description: "One sentence: which siblings under this scan_group you compared against and why this one is distinct." }
    FindingSummary:
      type: object
      required: [id, scan_id, repository_id, severity, location]
      properties:
        id:            { type: integer, format: int64 }
        scan_id:       { type: integer, format: int64 }
        repository_id: { type: integer, format: int64 }
        finding_id:    { type: string, description: Short id within the scan report (F1, F2, ...). }
        sinks:         { type: string, description: Comma-joined sink ids. }
        title:         { type: string }
        severity:      { type: string, enum: [Critical, High, Medium, Low] }
        severity_caps: { type: array, items: { type: string }, description: "Deterministic reasons that currently cap severity, derived from the latest reconciled verification controls and typed attacker prerequisites." }
        severity_calibration_incomplete: { type: boolean, description: "True when an unknown severity, unknown or unattempted prerequisite, unavailable control resolution, or unresolved control assessment prevented complete calibration; unknown inputs never lower severity." }
        status:        { type: string }
        cwe:           { type: string }
        location:      { type: string }
        vid:           { type: string, description: "Code identifier (VID-...) hashed from the sink code at the scanned commit; correlates findings across tools. Empty when not computed." }
        affected:      { type: string }
        reachability:  { type: string, enum: [reachable, harness_only, unclear] }
        quality_tier:  { type: string, enum: [high, low] }
        cve_id:        { type: string }
        ghsa_id:       { type: string, description: "GitHub Security Advisory id (GHSA-xxxx-xxxx-xxxx), set once the advisory is published." }
        cvss_vector:    { type: string, description: "CVSS v3.x base vector (3.0 or 3.1)." }
        cvss_score:     { type: number, description: "Base score derived from cvss_vector." }
        cvss_v4_vector: { type: string, description: "CVSS v4.0 base vector. Independent of cvss_vector; both may be populated." }
        cvss_v4_score:  { type: number, description: "Base score derived from cvss_v4_vector." }
        fix_version:   { type: string }
        fix_commit:    { type: string }
        resolution:    { type: string, enum: [fix, migrate, workaround, adopt, wontfix, ""] }
        assignee:      { type: string }
        exploited_in_wild: { type: string, enum: [yes, no, ""], description: "Analyst's call on whether this finding is being exploited in the wild. Empty means unknown. Automation never sets this." }
        exploited_in_wild_evidence: { type: string }
        mitigation:    { type: string, description: "Markdown body of operational mitigation guidance from the mitigate skill: workarounds, detection, optionally a semgrep rule (see mitigation_semgrep)." }
        mitigation_semgrep: { type: string, description: "Optional YAML semgrep rule from the mitigate skill. Empty when no rule was warranted." }
        dup_check:     { type: string, description: "Audit agent's one-sentence rationale for why this finding is distinct from the siblings already filed under the same scan_group. Empty for findings from skills that do not emit it." }
        novelty:       { type: string, enum: [unfixed, fixed, unclear, not_checked, ""], description: "Latest upstream novelty state. A touched file remains unclear until revalidate classifies the staged history." }
        production_viability: { type: string, enum: [VIABLE, NON_VIABLE, SAMPLE_OR_TEST, CONDITIONAL_VIABLE, ""], description: "Latest critic assessment of whether the finding can affect a real release build. Empty before assessment." }
        novelty_checked_commit: { type: string, description: "Repository HEAD compared with the finding's scanned commit for novelty." }
        novelty_checked_at: { type: string, format: date-time, nullable: true }
...